+++ API/controllers/service.controller.js
@@ -9,12 +9,6 @@ export const create = (req, res) => {
         });
         return;
     }
-    if (!req.body.libelle) {
-        res.status(400).send({
-            message: "Content can not be empty!"
-        });
-        return;
-    }
 
     const service = {
         libelle: req.body.libelle,
@@ -99,6 +93,13 @@ export const deleteOne = (req, res) => {
         });
         return;
     }
+    if (req.session.role !== "Administrateur") {
+        res.status(401).send({
+            message: "Vous devez être administrateur pour supprimer un service"
+        });
+        return;
+    }
+    
     const id = req.params.id;
 
     Service.destroy({
@@ -126,6 +127,12 @@ export const deleteAll = (req, res) => {
     if (!req.session.identifiant) {
         res.status(401).send({
             message: "Vous devez être connecté pour supprimer un service"
+        });
+        return;
+    }
+    if (req.session.role !== "Administrateur") {
+        res.status(401).send({
+            message: "Vous devez être administrateur pour supprimer un service"
         });
         return;
     }
