+++ API/controllers/standNext.controller.js
@@ -5,13 +5,7 @@ const Op = db.Sequelize.Op;
 export const create = (req, res) => {
     if (!req.session.identifiant) {
         res.status(401).send({
-            message: "Vous devez être connecté pour créer un artiste"
-        });
-        return;
-    }
-    if (!req.body.libelle) {
-        res.status(400).send({
-            message: "Content can not be empty!"
+            message: "Vous devez être connecté pour créer un stand"
         });
         return;
     }
@@ -67,6 +61,12 @@ export const findOne = (req, res) => {
 }
 
 export const update = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour modifier un stand"
+        });
+        return;
+    }
     const id = req.params.id;
 
     Stand.update(req.body, {
@@ -91,6 +91,18 @@ export const update = (req, res) => {
 }
 
 export const deleteOne = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer un stand"
+        });
+        return;
+    }
+    if (req.session.role !== "Administrateur") {
+        res.status(401).send({
+            message: "Vous devez être administrateur pour supprimer un stand"
+        });
+        return;
+    }
     const id = req.params.id;
 
     Stand.destroy({
@@ -115,6 +127,18 @@ export const deleteOne = (req, res) => {
 }
 
 export const deleteAll = (req, res) => {
+    if (!req.session.identifiant) {
+        res.status(401).send({
+            message: "Vous devez être connecté pour supprimer un stand"
+        });
+        return;
+    }
+    if (req.session.role !== "Administrateur") {
+        res.status(401).send({
+            message: "Vous devez être administrateur pour supprimer un stand"
+        });
+        return;
+    }
     Stand.destroy({
         where: {},
         truncate: false
